redpanda_to_aggregate/usage-api/requirements.txt
Flask==2.3.3
Flask-HTTPAuth==4.8.0
cassandra-driver==3.25.0
python-dotenv==1.0.0
Werkzeug==2.3.7